Rewrote the script get_network_graph.
The original script was using the DR lines in the top of the ibnetdiscover output in order to find and extract which LID belongs to which
switch/node. In my version of ibnetdiscover this method works for switches, but for some reason the lids for HCA ports in the DR lines are
set to zero. The updated script traverses the topology from the output of ibnetdiscover, and checks the comments on each line. In the
comments, the lids appear for each switch/hca-port. Also, the script uses regular expressions to extract values from matched strings based
on the ibnetdiscover output format.