This commit adds the infrastructure for a major new feature.
There are patterns that could benefit from allowing the user to pass a pattern-specific argument. A good example is the neighbor pattern that in order to run it for 2, 4, or 6 neighbors three different patterns had been defined.
A new command line argument option has been added, -a or --ptrnarg, that can be used for this precise reason. Pass pattern-specific arguments.
Eventually, instead of having three patterns for the 2, 4 and 6 neighbor pattern, the user can use a generic "neighbor" pattern and choose the number of communicating neighbors with the --ptrnarg option. As a bonus, we can now use the neighbor pattern with an arbitrary number of communicating neighbors.

Because the pattern-specific arguments may be of different types for different patterns (strings, int, double etc), a new struct has been added, cmdargs_t, that holds an instance of the "gengetopt_args_info args_info", and a pointer to a void ptrnarg.
I also added a new cpp file, cmdline_sanity.cpp, that has functions to perform more-than-basic sanity checks (some sanity checks are already provided by the code that is generated by gengetopts) for the command line arguments provided by the user.
The newly added function perform_sanity_checks_in_args is responsible to check if the user is using the --ptrnarg correctly for different patterns. Moreover, pattern-specific arguments will be converted to the necessary format (for example for the neighbor pattern, the arguments will be converted to integers) and the void pointer cmdargs->ptrnarg will be pointed to the corresponding data. At last, the function genptrn_by_name has been modified to be able to accept the void ptrnargs, cast the void to the necessary format, and call the corresponding pattern function.

Placeholder for a new "receivers" pattern has been added, but the pattern hasn't been implemented yet.

Now that the basic infrastructure is in place, as a next step I will remove options --ptrnfst and --ptrnsec and whenever the user wants to use ptrnvsptrn pattern, the first and second patterns will have to be provided by the --ptrnarg option in a string that will be looking like this: 'ptrn1:ptrnarg1,ptrn2:ptrnarg2', where the ptrnarg1 and ptrnarg2 will be optional.

Additional changelog for this commit:
* The Makefile has been updated and the simulator compiles fine.
* Some patterns that were not defined in the pattern_generator.hpp file have now been added.
* In the driver.cpp file wherever the args_info was used, it has now been replaced with the cmdargs.args_info.