The receivers pattern accepts an additional argument. The chance_to_not_communicate_at_all (or chance_factor_2).
The 'chance_factor_2' arg is another optional percentage (accepts values between 0.0 and 1.0) and
defines the chance that if a chosen source node is decided that will not communicate with a
receiver (based on chance_factor_1), there is a chance that it will stay idle (i.e. not communicate
at all with any other node). If the chance_factor_1 is set to 1.0, the chance_factor_2 will have no
effect in the experiment. The chance_factor_2 is set to 0.0 by default, i.e. there are no idle nodes.